mirroring branch component to plans component
